feat: Use Main class instead
